処理を責務ごとに分割
